Moved DeviceKeys to a separate file, renamed Device to IDevice
