+++ Project-Aurora/Project-Aurora/Devices/Device.cs
@@ -1,4 +1,4 @@
-﻿using System.Collections.Generic;
+using System.Collections.Generic;
 using System.ComponentModel;
 using System.Drawing;
 using System.Threading;
@@ -6,1308 +6,6 @@ using System.Threading.Tasks;
 
 namespace Aurora.Devices
 {
-    /// <summary>
-    /// Enum definition, representing everysingle supported device key
-    /// </summary>
-    public enum DeviceKeys
-    {
-        /// <summary>
-        /// Peripheral Device
-        /// <note type="note">Setting this key will make entire peripheral device one color</note>
-        /// </summary>
-        [Description("All Peripheral Devices")]
-        Peripheral = 0,
-
-        /// <summary>
-        /// Escape key
-        /// </summary>
-        [Description("Escape")]
-        ESC = 1,
-
-        /// <summary>
-        /// F1 key
-        /// </summary>
-        [Description("F1")]
-        F1 = 2,
-
-        /// <summary>
-        /// F2 key
-        /// </summary>
-        [Description("F2")]
-        F2 = 3,
-
-        /// <summary>
-        /// F3 key
-        /// </summary>
-        [Description("F3")]
-        F3 = 4,
-
-        /// <summary>
-        /// F4 key
-        /// </summary>
-        [Description("F4")]
-        F4 = 5,
-
-        /// <summary>
-        /// F5 key
-        /// </summary>
-        [Description("F5")]
-        F5 = 6,
-
-        /// <summary>
-        /// F6 key
-        /// </summary>
-        [Description("F6")]
-        F6 = 7,
-
-        /// <summary>
-        /// F7 key
-        /// </summary>
-        [Description("F7")]
-        F7 = 8,
-
-        /// <summary>
-        /// F8 key
-        /// </summary>
-        [Description("F8")]
-        F8 = 9,
-
-        /// <summary>
-        /// F9 key
-        /// </summary>
-        [Description("F9")]
-        F9 = 10,
-
-        /// <summary>
-        /// F10 key
-        /// </summary>
-        [Description("F10")]
-        F10 = 11,
-
-        /// <summary>
-        /// F11 key
-        /// </summary>
-        [Description("F11")]
-        F11 = 12,
-
-        /// <summary>
-        /// F12 key
-        /// </summary>
-        [Description("F12")]
-        F12 = 13,
-
-        /// <summary>
-        /// Print Screen key
-        /// </summary>
-        [Description("Print Screen")]
-        PRINT_SCREEN = 14,
-
-        /// <summary>
-        /// Scroll Lock key
-        /// </summary>
-        [Description("Scroll Lock")]
-        SCROLL_LOCK = 15,
-
-        /// <summary>
-        /// Pause/Break key
-        /// </summary>
-        [Description("Pause")]
-        PAUSE_BREAK = 16,
-
-
-        /// <summary>
-        /// Half/Full width (Japanese layout) key
-        /// </summary>
-        [Description("Half/Full width")]
-        JPN_HALFFULLWIDTH = 152,
-
-        /// <summary>
-        /// OEM 5 key
-        /// </summary>
-        [Description("OEM 5")]
-        OEM5 = 156,
-
-        /// <summary>
-        /// Tilde key
-        /// </summary>
-        [Description("Tilde")]
-        TILDE = 17,
-        
-        /// <summary>
-        /// One key
-        /// </summary>
-        [Description("1")]
-        ONE = 18,
-
-        /// <summary>
-        /// Two key
-        /// </summary>
-        [Description("2")]
-        TWO = 19,
-
-        /// <summary>
-        /// Three key
-        /// </summary>
-        [Description("3")]
-        THREE = 20,
-
-        /// <summary>
-        /// Four key
-        /// </summary>
-        [Description("4")]
-        FOUR = 21,
-
-        /// <summary>
-        /// Five key
-        /// </summary>
-        [Description("5")]
-        FIVE = 22,
-
-        /// <summary>
-        /// Six key
-        /// </summary>
-        [Description("6")]
-        SIX = 23,
-
-        /// <summary>
-        /// Seven key
-        /// </summary>
-        [Description("7")]
-        SEVEN = 24,
-
-        /// <summary>
-        /// Eight key
-        /// </summary>
-        [Description("8")]
-        EIGHT = 25,
-
-        /// <summary>
-        /// Nine key
-        /// </summary>
-        [Description("9")]
-        NINE = 26,
-
-        /// <summary>
-        /// Zero key
-        /// </summary>
-        [Description("0")]
-        ZERO = 27,
-
-        /// <summary>
-        /// Minus key
-        /// </summary>
-        [Description("-")]
-        MINUS = 28,
-
-        /// <summary>
-        /// Equals key
-        /// </summary>
-        [Description("=")]
-        EQUALS = 29,
-
-        /// <summary>
-        /// OEM 6 key
-        /// </summary>
-        [Description("OEM 6")]
-        OEM6 = 169,
-
-        /// <summary>
-        /// Backspace key
-        /// </summary>
-        [Description("Backspace")]
-        BACKSPACE = 30,
-
-        /// <summary>
-        /// Insert key
-        /// </summary>
-        [Description("Insert")]
-        INSERT = 31,
-
-        /// <summary>
-        /// Home key
-        /// </summary>
-        [Description("Home")]
-        HOME = 32,
-
-        /// <summary>
-        /// Page up key
-        /// </summary>
-        [Description("Page Up")]
-        PAGE_UP = 33,
-
-        /// <summary>
-        /// Numpad Lock key
-        /// </summary>
-        [Description("Numpad Lock")]
-        NUM_LOCK = 34,
-
-        /// <summary>
-        /// Numpad divide key
-        /// </summary>
-        [Description("Numpad /")]
-        NUM_SLASH = 35,
-
-        /// <summary>
-        /// Numpad multiply key
-        /// </summary>
-        [Description("Numpad *")]
-        NUM_ASTERISK = 36,
-
-        /// <summary>
-        /// Numpad minus key
-        /// </summary>
-        [Description("Numpad -")]
-        NUM_MINUS = 37,
-
-
-        /// <summary>
-        /// Tab key
-        /// </summary>
-        [Description("Tab")]
-        TAB = 38,
-
-        /// <summary>
-        /// Q key
-        /// </summary>
-        [Description("Q")]
-        Q = 39,
-
-        /// <summary>
-        /// W key
-        /// </summary>
-        [Description("W")]
-        W = 40,
-
-        /// <summary>
-        /// E key
-        /// </summary>
-        [Description("E")]
-        E = 41,
-
-        /// <summary>
-        /// R key
-        /// </summary>
-        [Description("R")]
-        R = 42,
-
-        /// <summary>
-        /// T key
-        /// </summary>
-        [Description("T")]
-        T = 43,
-
-        /// <summary>
-        /// Y key
-        /// </summary>
-        [Description("Y")]
-        Y = 44,
-
-        /// <summary>
-        /// U key
-        /// </summary>
-        [Description("U")]
-        U = 45,
-
-        /// <summary>
-        /// I key
-        /// </summary>
-        [Description("I")]
-        I = 46,
-
-        /// <summary>
-        /// O key
-        /// </summary>
-        [Description("O")]
-        O = 47,
-
-        /// <summary>
-        /// P key
-        /// </summary>
-        [Description("P")]
-        P = 48,
-
-        /// <summary>
-        /// OEM 1 key
-        /// </summary>
-        [Description("OEM 1")]
-        OEM1 = 170,
-
-        /// <summary>
-        /// Open Bracket key
-        /// </summary>
-        [Description("{")]
-        OPEN_BRACKET = 49,
-
-        /// <summary>
-        /// OEM Plus key
-        /// </summary>
-        [Description("OEM Plus")]
-        OEMPlus = 171,
-
-        /// <summary>
-        /// Close Bracket key
-        /// </summary>
-        [Description("}")]
-        CLOSE_BRACKET = 50,
-
-        /// <summary>
-        /// Backslash key
-        /// </summary>
-        [Description("\\")]
-        BACKSLASH = 51,
-
-        /// <summary>
-        /// Delete key
-        /// </summary>
-        [Description("Delete")]
-        DELETE = 52,
-
-        /// <summary>
-        /// End key
-        /// </summary>
-        [Description("End")]
-        END = 53,
-
-        /// <summary>
-        /// Page down key
-        /// </summary>
-        [Description("Page Down")]
-        PAGE_DOWN = 54,
-
-        /// <summary>
-        /// Numpad seven key
-        /// </summary>
-        [Description("Numpad 7")]
-        NUM_SEVEN = 55,
-
-        /// <summary>
-        /// Numpad eight key
-        /// </summary>
-        [Description("Numpad 8")]
-        NUM_EIGHT = 56,
-
-        /// <summary>
-        /// Numpad nine key
-        /// </summary>
-        [Description("Numpad 9")]
-        NUM_NINE = 57,
-
-        /// <summary>
-        /// Numpad add key
-        /// </summary>
-        [Description("Numpad +")]
-        NUM_PLUS = 58,
-
-
-        /// <summary>
-        /// Caps Lock key
-        /// </summary>
-        [Description("Caps Lock")]
-        CAPS_LOCK = 59,
-
-        /// <summary>
-        /// A key
-        /// </summary>
-        [Description("A")]
-        A = 60,
-
-        /// <summary>
-        /// S key
-        /// </summary>
-        [Description("S")]
-        S = 61,
-
-        /// <summary>
-        /// D key
-        /// </summary>
-        [Description("D")]
-        D = 62,
-
-        /// <summary>
-        /// F key
-        /// </summary>
-        [Description("F")]
-        F = 63,
-
-        /// <summary>
-        /// G key
-        /// </summary>
-        [Description("G")]
-        G = 64,
-
-        /// <summary>
-        /// H key
-        /// </summary>
-        [Description("H")]
-        H = 65,
-
-        /// <summary>
-        /// J key
-        /// </summary>
-        [Description("J")]
-        J = 66,
-
-        /// <summary>
-        /// K key
-        /// </summary>
-        [Description("K")]
-        K = 67,
-
-        /// <summary>
-        /// L key
-        /// </summary>
-        [Description("L")]
-        L = 68,
-
-        /// <summary>
-        /// OEM Tilde key
-        /// </summary>
-        [Description("OEM Tilde")]
-        OEMTilde = 157,
-
-        /// <summary>
-        /// Semicolon key
-        /// </summary>
-        [Description("Semicolon")]
-        SEMICOLON = 69,
-
-        /// <summary>
-        /// Apostrophe key
-        /// </summary>
-        [Description("Apostrophe")]
-        APOSTROPHE = 70,
-
-        /// <summary>
-        /// Hashtag key
-        /// </summary>
-        [Description("#")]
-        HASHTAG = 71,
-
-        /// <summary>
-        /// Enter key
-        /// </summary>
-        [Description("Enter")]
-        ENTER = 72,
-
-        /// <summary>
-        /// Numpad four key
-        /// </summary>
-        [Description("Numpad 4")]
-        NUM_FOUR = 73,
-
-        /// <summary>
-        /// Numpad five key
-        /// </summary>
-        [Description("Numpad 5")]
-        NUM_FIVE = 74,
-
-        /// <summary>
-        /// Numpad six key
-        /// </summary>
-        [Description("Numpad 6")]
-        NUM_SIX = 75,
-
-
-        /// <summary>
-        /// Left Shift key
-        /// </summary>
-        [Description("Left Shift")]
-        LEFT_SHIFT = 76,
-
-        /// <summary>
-        /// Non-US Backslash key
-        /// </summary>
-        [Description("Non-US Backslash")]
-        BACKSLASH_UK = 77,
-
-        /// <summary>
-        /// Z key
-        /// </summary>
-        [Description("Z")]
-        Z = 78,
-
-        /// <summary>
-        /// X key
-        /// </summary>
-        [Description("X")]
-        X = 79,
-
-        /// <summary>
-        /// C key
-        /// </summary>
-        [Description("C")]
-        C = 80,
-
-        /// <summary>
-        /// V key
-        /// </summary>
-        [Description("V")]
-        V = 81,
-
-        /// <summary>
-        /// B key
-        /// </summary>
-        [Description("B")]
-        B = 82,
-
-        /// <summary>
-        /// N key
-        /// </summary>
-        [Description("N")]
-        N = 83,
-
-        /// <summary>
-        /// M key
-        /// </summary>
-        [Description("M")]
-        M = 84,
-
-        /// <summary>
-        /// Comma key
-        /// </summary>
-        [Description("Comma")]
-        COMMA = 85,
-
-        /// <summary>
-        /// Period key
-        /// </summary>
-        [Description("Period")]
-        PERIOD = 86,
-
-        /// <summary>
-        /// Forward Slash key
-        /// </summary>
-        [Description("Forward Slash")]
-        FORWARD_SLASH = 87,
-
-        /// <summary>
-        /// OEM 8 key
-        /// </summary>
-        [Description("OEM 8")]
-        OEM8 = 158,
-
-        /// <summary>
-        /// OEM 102 key
-        /// </summary>
-        [Description("OEM 102")]
-        OEM102 = 159,
-
-        /// <summary>
-        /// Right Shift key
-        /// </summary>
-        [Description("Right Shift")]
-        RIGHT_SHIFT = 88,
-
-        /// <summary>
-        /// Arrow Up key
-        /// </summary>
-        [Description("Arrow Up")]
-        ARROW_UP = 89,
-
-        /// <summary>
-        /// Numpad one key
-        /// </summary>
-        [Description("Numpad 1")]
-        NUM_ONE = 90,
-
-        /// <summary>
-        /// Numpad two key
-        /// </summary>
-        [Description("Numpad 2")]
-        NUM_TWO = 91,
-
-        /// <summary>
-        /// Numpad three key
-        /// </summary>
-        [Description("Numpad 3")]
-        NUM_THREE = 92,
-
-        /// <summary>
-        /// Numpad enter key
-        /// </summary>
-        [Description("Numpad Enter")]
-        NUM_ENTER = 93,
-
-
-        /// <summary>
-        /// Left Control key
-        /// </summary>
-        [Description("Left Control")]
-        LEFT_CONTROL = 94,
-
-        /// <summary>
-        /// Left Windows key
-        /// </summary>
-        [Description("Left Windows Key")]
-        LEFT_WINDOWS = 95,
-
-        /// <summary>
-        /// Left Alt key
-        /// </summary>
-        [Description("Left Alt")]
-        LEFT_ALT = 96,
-
-        /// <summary>
-        /// Non-conversion (Japanese layout) key
-        /// </summary>
-        [Description("Non-conversion")]
-        JPN_MUHENKAN = 153,
-
-        /// <summary>
-        /// Spacebar key
-        /// </summary>
-        [Description("Spacebar")]
-        SPACE = 97,
-
-        /// <summary>
-        /// Conversion (Japanese layout) key
-        /// </summary>
-        [Description("Conversion")]
-        JPN_HENKAN = 154,
-
-        /// <summary>
-        /// Hiragana/Katakana (Japanese layout) key
-        /// </summary>
-        [Description("Hiragana/Katakana")]
-        JPN_HIRAGANA_KATAKANA = 155,
-
-        /// <summary>
-        /// Right Alt key
-        /// </summary>
-        [Description("Right Alt")]
-        RIGHT_ALT = 98,
-
-        /// <summary>
-        /// Right Windows key
-        /// </summary>
-        [Description("Right Windows Key")]
-        RIGHT_WINDOWS = 99,
-
-        /// <summary>
-        /// Application Select key
-        /// </summary>
-        [Description("Application Select Key")]
-        APPLICATION_SELECT = 100,
-
-        /// <summary>
-        /// Right Control key
-        /// </summary>
-        [Description("Right Control")]
-        RIGHT_CONTROL = 101,
-
-        /// <summary>
-        /// Arrow Left key
-        /// </summary>
-        [Description("Arrow Left")]
-        ARROW_LEFT = 102,
-
-        /// <summary>
-        /// Arrow Down key
-        /// </summary>
-        [Description("Arrow Down")]
-        ARROW_DOWN = 103,
-
-        /// <summary>
-        /// Arrow Right key
-        /// </summary>
-        [Description("Arrow Right")]
-        ARROW_RIGHT = 104,
-
-        /// <summary>
-        /// Numpad zero key
-        /// </summary>
-        [Description("Numpad 0")]
-        NUM_ZERO = 105,
-
-        /// <summary>
-        /// Numpad period key
-        /// </summary>
-        [Description("Numpad Period")]
-        NUM_PERIOD = 106,
-
-
-        /// <summary>
-        /// Function key
-        /// </summary>
-        [Description("FN Key")]
-        FN_Key = 107,
-
-
-        /// <summary>
-        /// Macrokey 1 key
-        /// </summary>
-        [Description("G1")]
-        G1 = 108,
-
-        /// <summary>
-        /// Macrokey 2 key
-        /// </summary>
-        [Description("G2")]
-        G2 = 109,
-
-        /// <summary>
-        /// Macrokey 3 key
-        /// </summary>
-        [Description("G3")]
-        G3 = 110,
-
-        /// <summary>
-        /// Macrokey 4 key
-        /// </summary>
-        [Description("G4")]
-        G4 = 111,
-
-        /// <summary>
-        /// Macrokey 5 key
-        /// </summary>
-        [Description("G5")]
-        G5 = 112,
-
-        /// <summary>
-        /// Macrokey 6 key
-        /// </summary>
-        [Description("G6")]
-        G6 = 113,
-
-        /// <summary>
-        /// Macrokey 7 key
-        /// </summary>
-        [Description("G7")]
-        G7 = 114,
-
-        /// <summary>
-        /// Macrokey 8 key
-        /// </summary>
-        [Description("G8")]
-        G8 = 115,
-
-        /// <summary>
-        /// Macrokey 9 key
-        /// </summary>
-        [Description("G9")]
-        G9 = 116,
-
-        /// <summary>
-        /// Macrokey 10 key
-        /// </summary>
-        [Description("G10")]
-        G10 = 117,
-
-        /// <summary>
-        /// Macrokey 11 key
-        /// </summary>
-        [Description("G11")]
-        G11 = 118,
-
-        /// <summary>
-        /// Macrokey 12 key
-        /// </summary>
-        [Description("G12")]
-        G12 = 119,
-
-        /// <summary>
-        /// Macrokey 13 key
-        /// </summary>
-        [Description("G13")]
-        G13 = 120,
-
-        /// <summary>
-        /// Macrokey 14 key
-        /// </summary>
-        [Description("G14")]
-        G14 = 121,
-
-        /// <summary>
-        /// Macrokey 15 key
-        /// </summary>
-        [Description("G15")]
-        G15 = 122,
-
-        /// <summary>
-        /// Macrokey 16 key
-        /// </summary>
-        [Description("G16")]
-        G16 = 123,
-
-        /// <summary>
-        /// Macrokey 17 key
-        /// </summary>
-        [Description("G17")]
-        G17 = 124,
-
-        /// <summary>
-        /// Macrokey 18 key
-        /// </summary>
-        [Description("G18")]
-        G18 = 125,
-
-        /// <summary>
-        /// Macrokey 19 key
-        /// </summary>
-        [Description("G19")]
-        G19 = 126,
-
-        /// <summary>
-        /// Macrokey 20 key
-        /// </summary>
-        [Description("G20")]
-        G20 = 127,
-
-
-        /// <summary>
-        /// Brand Logo
-        /// </summary>
-        [Description("Brand Logo")]
-        LOGO = 128,
-
-        /// <summary>
-        /// Brand Logo #2
-        /// </summary>
-        [Description("Brand Logo #2")]
-        LOGO2 = 129,
-
-        /// <summary>
-        /// Brand Logo #3
-        /// </summary>
-        [Description("Brand Logo #3")]
-        LOGO3 = 130,
-
-        /// <summary>
-        /// Brightness Switch
-        /// </summary>
-        [Description("Brightness Switch")]
-        BRIGHTNESS_SWITCH = 131,
-
-        /// <summary>
-        /// Lock Switch
-        /// </summary>
-        [Description("Lock Switch")]
-        LOCK_SWITCH = 132,
-
-
-        /// <summary>
-        /// Multimedia Play/Pause
-        /// </summary>
-        [Description("Media Play/Pause")]
-        MEDIA_PLAY_PAUSE = 133,
-
-        /// <summary>
-        /// Multimedia Play
-        /// </summary>
-        [Description("Media Play")]
-        MEDIA_PLAY = 134,
-
-        /// <summary>
-        /// Multimedia Pause
-        /// </summary>
-        [Description("Media Pause")]
-        MEDIA_PAUSE = 135,
-
-        /// <summary>
-        /// Multimedia Stop
-        /// </summary>
-        [Description("Media Stop")]
-        MEDIA_STOP = 136,
-
-        /// <summary>
-        /// Multimedia Previous
-        /// </summary>
-        [Description("Media Previous")]
-        MEDIA_PREVIOUS = 137,
-
-        /// <summary>
-        /// Multimedia Next
-        /// </summary>
-        [Description("Media Next")]
-        MEDIA_NEXT = 138,
-
-
-        /// <summary>
-        /// Volume Mute
-        /// </summary>
-        [Description("Volume Mute")]
-        VOLUME_MUTE = 139,
-
-        /// <summary>
-        /// Volume Down
-        /// </summary>
-        [Description("Volume Down")]
-        VOLUME_DOWN = 140,
-
-        /// <summary>
-        /// Volume Up
-        /// </summary>
-        [Description("Volume Up")]
-        VOLUME_UP = 141,
-
-
-        /// <summary>
-        /// Additional Light 1
-        /// </summary>
-        [Description("Additional Light 1")]
-        ADDITIONALLIGHT1 = 142,
-
-        /// <summary>
-        /// Additional Light 2
-        /// </summary>
-        [Description("Additional Light 2")]
-        ADDITIONALLIGHT2 = 143,
-
-        /// <summary>
-        /// Additional Light 3
-        /// </summary>
-        [Description("Additional Light 3")]
-        ADDITIONALLIGHT3 = 144,
-
-        /// <summary>
-        /// Additional Light 4
-        /// </summary>
-        [Description("Additional Light 4")]
-        ADDITIONALLIGHT4 = 145,
-
-        /// <summary>
-        /// Additional Light 5
-        /// </summary>
-        [Description("Additional Light 5")]
-        ADDITIONALLIGHT5 = 146,
-
-        /// <summary>
-        /// Additional Light 6
-        /// </summary>
-        [Description("Additional Light 6")]
-        ADDITIONALLIGHT6 = 147,
-
-        /// <summary>
-        /// Additional Light 7
-        /// </summary>
-        [Description("Additional Light 7")]
-        ADDITIONALLIGHT7 = 148,
-
-        /// <summary>
-        /// Additional Light 8
-        /// </summary>
-        [Description("Additional Light 8")]
-        ADDITIONALLIGHT8 = 149,
-
-        /// <summary>
-        /// Additional Light 9
-        /// </summary>
-        [Description("Additional Light 9")]
-        ADDITIONALLIGHT9 = 150,
-
-        /// <summary>
-        /// Additional Light 10
-        /// </summary>
-        [Description("Additional Light 10")]
-        ADDITIONALLIGHT10 = 151,
-
-        /// <summary>
-        /// Peripheral Logo
-        /// </summary>
-        [Description("Peripheral Logo")]
-        Peripheral_Logo = 160,
-
-        /// <summary>
-        /// Peripheral Scroll Wheel
-        /// </summary>
-        [Description("Peripheral Scroll Wheel")]
-        Peripheral_ScrollWheel = 161,
-
-        /// <summary>
-        /// Peripheral Front-facing lights
-        /// </summary>
-        [Description("Peripheral Front Lights")]
-        Peripheral_FrontLight = 162,
-
-        /// <summary>
-        /// Profile key 1
-        /// </summary>
-        [Description("Profile Key 1")]
-        Profile_Key1 = 163,
-
-        /// <summary>
-        /// Profile key 2
-        /// </summary>
-        [Description("Profile Key 2")]
-        Profile_Key2 = 164,
-
-        /// <summary>
-        /// Profile key 3
-        /// </summary>
-        [Description("Profile Key 3")]
-        Profile_Key3 = 165,
-
-        /// <summary>
-        /// Profile key 4
-        /// </summary>
-        [Description("Profile Key 4")]
-        Profile_Key4 = 166,
-
-        /// <summary>
-        /// Profile key 5
-        /// </summary>
-        [Description("Profile Key 5")]
-        Profile_Key5 = 167,
-
-        /// <summary>
-        /// Profile key 6
-        /// </summary>
-        [Description("Profile Key 6")]
-        Profile_Key6 = 168,
-
-        /// <summary>
-        /// Numpad 00
-        /// </summary>
-        [Description("Numpad 00")]
-        NUM_ZEROZERO = 169,
-
-        /// <summary>
-        /// Macrokey 0 key
-        /// </summary>
-        [Description("G0")]
-        G0 = 170,
-
-        /// <summary>
-        /// Macrokey 0 key
-        /// </summary>
-        [Description("Left FN")]
-        LEFT_FN = 171,
-
-        /// <summary>
-        /// Additional Light 11
-        /// </summary>
-        [Description("Additional Light 11")]
-        ADDITIONALLIGHT11 = 172,
-
-        /// <summary>
-        /// Additional Light 12
-        /// </summary>
-        [Description("Additional Light 12")]
-        ADDITIONALLIGHT12 = 173,
-
-        /// <summary>
-        /// Additional Light 13
-        /// </summary>
-        [Description("Additional Light 13")]
-        ADDITIONALLIGHT13 = 174,
-
-        /// <summary>
-        /// Additional Light 14
-        /// </summary>
-        [Description("Additional Light 14")]
-        ADDITIONALLIGHT14 = 175,
-
-        /// <summary>
-        /// Additional Light 15
-        /// </summary>
-        [Description("Additional Light 15")]
-        ADDITIONALLIGHT15 = 176,
-
-        /// <summary>
-        /// Additional Light 16
-        /// </summary>
-        [Description("Additional Light 16")]
-        ADDITIONALLIGHT16 = 177,
-
-        /// <summary>
-        /// Additional Light 17
-        /// </summary>
-        [Description("Additional Light 17")]
-        ADDITIONALLIGHT17 = 178,
-
-        /// <summary>
-        /// Additional Light 18
-        /// </summary>
-        [Description("Additional Light 18")]
-        ADDITIONALLIGHT18 = 179,
-
-        /// <summary>
-        /// Additional Light 19
-        /// </summary>
-        [Description("Additional Light 19")]
-        ADDITIONALLIGHT19 = 180,
-
-        /// <summary>
-        /// Additional Light 20
-        /// </summary>
-        [Description("Additional Light 20")]
-        ADDITIONALLIGHT20 = 181,
-
-        /// <summary>
-        /// Additional Light 21
-        /// </summary>
-        [Description("Additional Light 21")]
-        ADDITIONALLIGHT21 = 182,
-
-        /// <summary>
-        /// Additional Light 22
-        /// </summary>
-        [Description("Additional Light 22")]
-        ADDITIONALLIGHT22 = 183,
-
-        /// <summary>
-        /// Additional Light 23
-        /// </summary>
-        [Description("Additional Light 23")]
-        ADDITIONALLIGHT23 = 184,
-
-        /// <summary>
-        /// Additional Light 24
-        /// </summary>
-        [Description("Additional Light 24")]
-        ADDITIONALLIGHT24 = 185,
-
-        /// <summary>
-        /// Additional Light 25
-        /// </summary>
-        [Description("Additional Light 25")]
-        ADDITIONALLIGHT25 = 186,
-
-        /// <summary>
-        /// Additional Light 26
-        /// </summary>
-        [Description("Additional Light 26")]
-        ADDITIONALLIGHT26 = 187,
-
-        /// <summary>
-        /// Additional Light 27
-        /// </summary>
-        [Description("Additional Light 27")]
-        ADDITIONALLIGHT27 = 188,
-        
-        /// <summary>
-        /// Additional Light 28
-        /// </summary>
-        [Description("Additional Light 28")]
-        ADDITIONALLIGHT28 = 189,
-
-        /// <summary>
-        /// Additional Light 29
-        /// </summary>
-        [Description("Additional Light 29")]
-        ADDITIONALLIGHT29 = 190,
-
-        /// <summary>
-        /// Additional Light 30
-        /// </summary>
-        [Description("Additional Light 30")]
-        ADDITIONALLIGHT30 = 191,
-
-        /// <summary>
-        /// Additional Light 31
-        /// </summary>
-        [Description("Additional Light 31")]
-        ADDITIONALLIGHT31 = 192,
-
-        /// <summary>
-        /// Additional Light 32
-        /// </summary>
-        [Description("Additional Light 32")]
-        ADDITIONALLIGHT32 = 193,
-
-        /// <summary>
-        /// Mousepad Light 1
-        /// </summary>
-        [Description("Mousepad Light 1")]
-        MOUSEPADLIGHT1 = 201,
-
-        /// <summary>
-        /// Mousepad Light 2
-        /// </summary>
-        [Description("Mousepad Light 2")]
-        MOUSEPADLIGHT2 = 202,
-
-        /// <summary>
-        /// Mousepad Light 1
-        /// </summary>
-        [Description("Mousepad Light 3")]
-        MOUSEPADLIGHT3 = 203,
-
-        /// <summary>
-        /// Mousepad Light 2
-        /// </summary>
-        [Description("Mousepad Light 4")]
-        MOUSEPADLIGHT4 = 204,
-
-        /// <summary>
-        /// Mousepad Light 1
-        /// </summary>
-        [Description("Mousepad Light 5")]
-        MOUSEPADLIGHT5 = 205,
-
-        /// <summary>
-        /// Mousepad Light 2
-        /// </summary>
-        [Description("Mousepad Light 6")]
-        MOUSEPADLIGHT6 = 206,
-
-        /// <summary>
-        /// Mousepad Light 1
-        /// </summary>
-        [Description("Mousepad Light 7")]
-        MOUSEPADLIGHT7 = 207,
-
-        /// <summary>
-        /// Mousepad Light 2
-        /// </summary>
-        [Description("Mousepad Light 8")]
-        MOUSEPADLIGHT8 = 208,
-
-        /// <summary>
-        /// Mousepad Light 1
-        /// </summary>
-        [Description("Mousepad Light 9")]
-        MOUSEPADLIGHT9 = 209,
-
-        /// <summary>
-        /// Mousepad Light 2
-        /// </summary>
-        [Description("Mousepad Light 10")]
-        MOUSEPADLIGHT10 = 210,
-
-        /// <summary>
-        /// Mousepad Light 2
-        /// </summary>
-        [Description("Mousepad Light 11")]
-        MOUSEPADLIGHT11 = 211,
-
-        /// <summary>
-        /// Mousepad Light 1
-        /// </summary>
-        [Description("Mousepad Light 12")]
-        MOUSEPADLIGHT12 = 212,
-
-        /// <summary>
-        /// Mousepad Light 2
-        /// </summary>
-        [Description("Mousepad Light 13")]
-        MOUSEPADLIGHT13 = 213,
-
-        /// <summary>
-        /// Mousepad Light 1
-        /// </summary>
-        [Description("Mousepad Light 14")]
-        MOUSEPADLIGHT14 = 214,
-
-        /// <summary>
-        /// Mousepad Light 2
-        /// </summary>
-        [Description("Mousepad Light 15")]
-        MOUSEPADLIGHT15 = 215,
-
-        ///<summary>
-        /// Calculator Key
-        /// </summary>
-        [Description("Calculator")]
-        CALC = 216,
-
-        /// <summary>
-        /// None
-        /// </summary>
-        [Description("None")]
-        NONE = -1,
-    };
-
     /// <summary>
     /// Struct representing color settings being sent to devices
     /// </summary>
@@ -1321,7 +19,7 @@ namespace Aurora.Devices
     /// <summary>
     /// An interface for a device class.
     /// </summary>
-    public interface Device
+    public interface IDevice
     {
         /// <summary>
         /// Gets registered variables by this device.
