+++ Project-Aurora/Project-Aurora/Devices/YeeLight/YeeLightDevice.cs
@@ -16,7 +16,7 @@ using YeeLightAPI.YeeLightExceptions;
 
 namespace Aurora.Devices.YeeLight
 {
-    public class YeeLightDevice : Device
+    public class YeeLightDevice : IDevice
     {
         private string devicename = "YeeLight";
